requirements.txt
ccxt>=4.2.14
pandas>=2.2.0
numpy>=1.26.4
google-genai>=0.1.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
joblib>=1.3.2
yfinance>=0.2.36
ta>=0.11.0
beautifulsoup4>=4.12.3
chromadb>=0.4.22
python-binance>=1.0.19
fastapi>=0.110.0
uvicorn>=0.27.0
aiosqlite>=0.20.0
aiohttp>=3.9.3
requests>=2.31.0
